+++ DatadogInternal/Sources/Models/RUM/RUMDataModels.swift
@@ -341,8 +341,19 @@ extension RUMAccount {
 
         // Encode dynamic properties:
         var dynamicContainer = encoder.container(keyedBy: DynamicCodingKey.self)
-        accountInfo.forEach { name, value in
-            dynamicContainer.encodeAttribute(AnyEncodable(value), forKey: DynamicCodingKey(name), attributeName: name, context: .accountInfo)
+        let shouldRecover = encoder.shouldRecoverAttributeFailures
+        try accountInfo.forEach { name, value in
+            if shouldRecover, StaticCodingKeys(stringValue: name) != nil {
+                return
+            }
+
+            try dynamicContainer.encodeAttribute(
+                AnyEncodable(value),
+                forKey: DynamicCodingKey(name),
+                attributeName: name,
+                context: .accountInfo,
+                shouldRecover: shouldRecover
+            )
         }
     }
 
@@ -2706,8 +2717,15 @@ extension RUMErrorEvent.FeatureFlags {
     public func encode(to encoder: Encoder) throws {
         // Encode dynamic properties:
         var dynamicContainer = encoder.container(keyedBy: DynamicCodingKey.self)
-        featureFlagsInfo.forEach { name, value in
-            dynamicContainer.encodeAttribute(AnyEncodable(value), forKey: DynamicCodingKey(name), attributeName: name, context: .custom)
+        let shouldRecover = encoder.shouldRecoverAttributeFailures
+        try featureFlagsInfo.forEach { name, value in
+            try dynamicContainer.encodeAttribute(
+                AnyEncodable(value),
+                forKey: DynamicCodingKey(name),
+                attributeName: name,
+                context: .custom,
+                shouldRecover: shouldRecover
+            )
         }
     }
 
@@ -2741,8 +2759,15 @@ extension RUMEventAttributes {
     public func encode(to encoder: Encoder) throws {
         // Encode dynamic properties:
         var dynamicContainer = encoder.container(keyedBy: DynamicCodingKey.self)
-        contextInfo.forEach { name, value in
-            dynamicContainer.encodeAttribute(AnyEncodable(value), forKey: DynamicCodingKey(name), attributeName: name, context: .custom)
+        let shouldRecover = encoder.shouldRecoverAttributeFailures
+        try contextInfo.forEach { name, value in
+            try dynamicContainer.encodeAttribute(
+                AnyEncodable(value),
+                forKey: DynamicCodingKey(name),
+                attributeName: name,
+                context: .custom,
+                shouldRecover: shouldRecover
+            )
         }
     }
 
@@ -4964,8 +4989,15 @@ extension RUMResourceEvent.Resource.Request.Headers {
     public func encode(to encoder: Encoder) throws {
         // Encode dynamic properties:
         var dynamicContainer = encoder.container(keyedBy: DynamicCodingKey.self)
-        headersInfo.forEach { name, value in
-            dynamicContainer.encodeAttribute(value, forKey: DynamicCodingKey(name), attributeName: name, context: .custom)
+        let shouldRecover = encoder.shouldRecoverAttributeFailures
+        try headersInfo.forEach { name, value in
+            try dynamicContainer.encodeAttribute(
+                value,
+                forKey: DynamicCodingKey(name),
+                attributeName: name,
+                context: .custom,
+                shouldRecover: shouldRecover
+            )
         }
     }
 
@@ -4984,8 +5016,15 @@ extension RUMResourceEvent.Resource.Response.Headers {
     public func encode(to encoder: Encoder) throws {
         // Encode dynamic properties:
         var dynamicContainer = encoder.container(keyedBy: DynamicCodingKey.self)
-        headersInfo.forEach { name, value in
-            dynamicContainer.encodeAttribute(value, forKey: DynamicCodingKey(name), attributeName: name, context: .custom)
+        let shouldRecover = encoder.shouldRecoverAttributeFailures
+        try headersInfo.forEach { name, value in
+            try dynamicContainer.encodeAttribute(
+                value,
+                forKey: DynamicCodingKey(name),
+                attributeName: name,
+                context: .custom,
+                shouldRecover: shouldRecover
+            )
         }
     }
 
@@ -5067,8 +5106,19 @@ extension RUMSyntheticsTest {
 
         // Encode dynamic properties:
         var dynamicContainer = encoder.container(keyedBy: DynamicCodingKey.self)
-        syntheticsInfo.forEach { name, value in
-            dynamicContainer.encodeAttribute(AnyEncodable(value), forKey: DynamicCodingKey(name), attributeName: name, context: .internal)
+        let shouldRecover = encoder.shouldRecoverAttributeFailures
+        try syntheticsInfo.forEach { name, value in
+            if shouldRecover, StaticCodingKeys(stringValue: name) != nil {
+                return
+            }
+
+            try dynamicContainer.encodeAttribute(
+                AnyEncodable(value),
+                forKey: DynamicCodingKey(name),
+                attributeName: name,
+                context: .internal,
+                shouldRecover: shouldRecover
+            )
         }
     }
 
@@ -6451,8 +6501,19 @@ extension RUMUser {
 
         // Encode dynamic properties:
         var dynamicContainer = encoder.container(keyedBy: DynamicCodingKey.self)
-        usrInfo.forEach { name, value in
-            dynamicContainer.encodeAttribute(AnyEncodable(value), forKey: DynamicCodingKey(name), attributeName: name, context: .userInfo)
+        let shouldRecover = encoder.shouldRecoverAttributeFailures
+        try usrInfo.forEach { name, value in
+            if shouldRecover, StaticCodingKeys(stringValue: name) != nil {
+                return
+            }
+
+            try dynamicContainer.encodeAttribute(
+                AnyEncodable(value),
+                forKey: DynamicCodingKey(name),
+                attributeName: name,
+                context: .userInfo,
+                shouldRecover: shouldRecover
+            )
         }
     }
 
@@ -8613,8 +8674,15 @@ extension RUMViewEvent.FeatureFlags {
     public func encode(to encoder: Encoder) throws {
         // Encode dynamic properties:
         var dynamicContainer = encoder.container(keyedBy: DynamicCodingKey.self)
-        featureFlagsInfo.forEach { name, value in
-            dynamicContainer.encodeAttribute(AnyEncodable(value), forKey: DynamicCodingKey(name), attributeName: name, context: .custom)
+        let shouldRecover = encoder.shouldRecoverAttributeFailures
+        try featureFlagsInfo.forEach { name, value in
+            try dynamicContainer.encodeAttribute(
+                AnyEncodable(value),
+                forKey: DynamicCodingKey(name),
+                attributeName: name,
+                context: .custom,
+                shouldRecover: shouldRecover
+            )
         }
     }
 
@@ -8633,8 +8701,15 @@ extension RUMViewEvent.View.CustomTimings {
     public func encode(to encoder: Encoder) throws {
         // Encode dynamic properties:
         var dynamicContainer = encoder.container(keyedBy: DynamicCodingKey.self)
-        customTimingsInfo.forEach { name, value in
-            dynamicContainer.encodeAttribute(value, forKey: DynamicCodingKey(name), attributeName: name, context: .custom)
+        let shouldRecover = encoder.shouldRecoverAttributeFailures
+        try customTimingsInfo.forEach { name, value in
+            try dynamicContainer.encodeAttribute(
+                value,
+                forKey: DynamicCodingKey(name),
+                attributeName: name,
+                context: .custom,
+                shouldRecover: shouldRecover
+            )
         }
     }
 
@@ -10787,8 +10862,15 @@ extension RUMViewUpdateEvent.FeatureFlags {
     public func encode(to encoder: Encoder) throws {
         // Encode dynamic properties:
         var dynamicContainer = encoder.container(keyedBy: DynamicCodingKey.self)
-        featureFlagsInfo.forEach { name, value in
-            dynamicContainer.encodeAttribute(AnyEncodable(value), forKey: DynamicCodingKey(name), attributeName: name, context: .custom)
+        let shouldRecover = encoder.shouldRecoverAttributeFailures
+        try featureFlagsInfo.forEach { name, value in
+            try dynamicContainer.encodeAttribute(
+                AnyEncodable(value),
+                forKey: DynamicCodingKey(name),
+                attributeName: name,
+                context: .custom,
+                shouldRecover: shouldRecover
+            )
         }
     }
 
@@ -10807,8 +10889,15 @@ extension RUMViewUpdateEvent.View.CustomTimings {
     public func encode(to encoder: Encoder) throws {
         // Encode dynamic properties:
         var dynamicContainer = encoder.container(keyedBy: DynamicCodingKey.self)
-        customTimingsInfo.forEach { name, value in
-            dynamicContainer.encodeAttribute(value, forKey: DynamicCodingKey(name), attributeName: name, context: .custom)
+        let shouldRecover = encoder.shouldRecoverAttributeFailures
+        try customTimingsInfo.forEach { name, value in
+            try dynamicContainer.encodeAttribute(
+                value,
+                forKey: DynamicCodingKey(name),
+                attributeName: name,
+                context: .custom,
+                shouldRecover: shouldRecover
+            )
         }
     }
 
@@ -13902,8 +13991,19 @@ extension TelemetryConfigurationEvent.Telemetry {
 
         // Encode dynamic properties:
         var dynamicContainer = encoder.container(keyedBy: DynamicCodingKey.self)
-        telemetryInfo.forEach { name, value in
-            dynamicContainer.encodeAttribute(AnyEncodable(value), forKey: DynamicCodingKey(name), attributeName: name, context: .internal)
+        let shouldRecover = encoder.shouldRecoverAttributeFailures
+        try telemetryInfo.forEach { name, value in
+            if shouldRecover, StaticCodingKeys(stringValue: name) != nil {
+                return
+            }
+
+            try dynamicContainer.encodeAttribute(
+                AnyEncodable(value),
+                forKey: DynamicCodingKey(name),
+                attributeName: name,
+                context: .internal,
+                shouldRecover: shouldRecover
+            )
         }
     }
 
@@ -13933,8 +14033,19 @@ extension TelemetryConfigurationEvent.Telemetry.Configuration.Plugins {
 
         // Encode dynamic properties:
         var dynamicContainer = encoder.container(keyedBy: DynamicCodingKey.self)
-        pluginsInfo.forEach { name, value in
-            dynamicContainer.encodeAttribute(AnyEncodable(value), forKey: DynamicCodingKey(name), attributeName: name, context: .custom)
+        let shouldRecover = encoder.shouldRecoverAttributeFailures
+        try pluginsInfo.forEach { name, value in
+            if shouldRecover, StaticCodingKeys(stringValue: name) != nil {
+                return
+            }
+
+            try dynamicContainer.encodeAttribute(
+                AnyEncodable(value),
+                forKey: DynamicCodingKey(name),
+                attributeName: name,
+                context: .custom,
+                shouldRecover: shouldRecover
+            )
         }
     }
 
@@ -14221,8 +14332,19 @@ extension TelemetryDebugEvent.Telemetry {
 
         // Encode dynamic properties:
         var dynamicContainer = encoder.container(keyedBy: DynamicCodingKey.self)
-        telemetryInfo.forEach { name, value in
-            dynamicContainer.encodeAttribute(AnyEncodable(value), forKey: DynamicCodingKey(name), attributeName: name, context: .internal)
+        let shouldRecover = encoder.shouldRecoverAttributeFailures
+        try telemetryInfo.forEach { name, value in
+            if shouldRecover, StaticCodingKeys(stringValue: name) != nil {
+                return
+            }
+
+            try dynamicContainer.encodeAttribute(
+                AnyEncodable(value),
+                forKey: DynamicCodingKey(name),
+                attributeName: name,
+                context: .internal,
+                shouldRecover: shouldRecover
+            )
         }
     }
 
@@ -14546,8 +14668,19 @@ extension TelemetryErrorEvent.Telemetry {
 
         // Encode dynamic properties:
         var dynamicContainer = encoder.container(keyedBy: DynamicCodingKey.self)
-        telemetryInfo.forEach { name, value in
-            dynamicContainer.encodeAttribute(AnyEncodable(value), forKey: DynamicCodingKey(name), attributeName: name, context: .internal)
+        let shouldRecover = encoder.shouldRecoverAttributeFailures
+        try telemetryInfo.forEach { name, value in
+            if shouldRecover, StaticCodingKeys(stringValue: name) != nil {
+                return
+            }
+
+            try dynamicContainer.encodeAttribute(
+                AnyEncodable(value),
+                forKey: DynamicCodingKey(name),
+                attributeName: name,
+                context: .internal,
+                shouldRecover: shouldRecover
+            )
         }
     }
 
@@ -15547,8 +15680,19 @@ extension TelemetryUsageEvent.Telemetry {
 
         // Encode dynamic properties:
         var dynamicContainer = encoder.container(keyedBy: DynamicCodingKey.self)
-        telemetryInfo.forEach { name, value in
-            dynamicContainer.encodeAttribute(AnyEncodable(value), forKey: DynamicCodingKey(name), attributeName: name, context: .internal)
+        let shouldRecover = encoder.shouldRecoverAttributeFailures
+        try telemetryInfo.forEach { name, value in
+            if shouldRecover, StaticCodingKeys(stringValue: name) != nil {
+                return
+            }
+
+            try dynamicContainer.encodeAttribute(
+                AnyEncodable(value),
+                forKey: DynamicCodingKey(name),
+                attributeName: name,
+                context: .internal,
+                shouldRecover: shouldRecover
+            )
         }
     }