+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - [FEATURE] Add an experimental Core Animation recording pipeline for Session Replay, available through the `compositionTreeRecording` feature flag. See [#3127][]
 - [FEATURE] Add `disallowList` to `RUM.Configuration.URLSessionTracking` to exclude URLs from automatic RUM resource tracking, with `*` wildcard support. [#3097][]
 - [IMPROVEMENT] Forward `local_cache_hit` signal on RUM resources [#3074][]
+- [FIX] Fix `EXC_BREAKPOINT` crash when a log or RUM attribute's `encode(to:)` throws after partially encoding a value. [#3134][]
 
 # 3.15.0 / 05-08-2026
 
@@ -1225,6 +1226,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#3074]: https://github.com/DataDog/dd-sdk-ios/pull/3074
 [#3127]: https://github.com/DataDog/dd-sdk-ios/pull/3127
 [#3097]: https://github.com/DataDog/dd-sdk-ios/pull/3097
+[#3134]: https://github.com/DataDog/dd-sdk-ios/pull/3134
 
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin
@@ -1264,3 +1266,4 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [@blimmer]: https://github.com/blimmer
 [@thedavidharris]: https://github.com/thedavidharris
 [@noremac]: https://github.com/noremac
+[@saladdays831]: https://github.com/saladdays831